pyproject.toml
[tool.poetry]
name = "playerservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standart"], version = "^0.115.8"}
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
aiosqlite = "^0.21.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
